Move method injection tests over to functional/
